V6/requirements.txt
fastapi
uvicorn
llama-cpp-python
redis
requests
psutil